agent-a-news-engine/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
apscheduler==3.10.4
httpx==0.27.0
feedparser==6.0.11
supabase==2.7.0
openai==1.40.0
google-generativeai==0.7.2
anthropic==0.34.0
python-dotenv==1.0.1
